api-chatgpt/requirements.txt
openai
python-dotenv
fastapi
uvicorn
python-multipart
pydantic